backend/requirements.txt
# FastAPI Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
aiosqlite==0.19.0
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client
httpx==0.26.0

# File Processing
python-multipart==0.0.6
aiofiles==23.2.1

# Encryption
cryptography==41.0.7

# Utilities
python-dotenv==1.0.0
